+++ pyArango/tests/tests.py
@@ -85,6 +85,19 @@ class pyArangoTests(unittest.TestCase):
         self.assertEqual(res, nbUsers)
 
     # @unittest.skip("stand by")
+    def test_bulkSave_dict(self) :
+        collection = self.db.createCollection(name = "lops")
+        nbUsers = 100
+        docs = []
+        for i in range(nbUsers) :
+            doc = {}
+            doc["name"] = "Tesla-%d" % i
+            docs.append(doc)
+
+        res = collection.bulkSave(docs)
+        self.assertEqual(res, nbUsers)
+
+    # @unittest.skip("stand by")
     def test_collection_create_delete(self) :
         col = self.db.createCollection(name = "to_be_erased")
         self.assertTrue(self.db.hasCollection("to_be_erased"))