+++ CHANGELOG.rst
@@ -8,6 +8,7 @@
 * stastd support thx to: @dothebart
 * properties definition in schema
 * AQL errors now come with prints and line numbers for everyone's convenience
+* Bulk save for Document objects and dicts
 
 1.3.1
 =====